fix(api): ensure critical points are acceptable for a 96 channel
We were using the original calculations for 8 channel pipttes in the Y direction for both XY_CENTER
and FRONT_NOZZLE. This logic works because the Y element equals the number of channels. That is no
longer the case for the 96 channels so we need to temporarily modify the calculation so that the
critical point calculation is correct for the 96 channel AND 8 channel. The long term fix is to add
critical point information for all pipette nozzles as individual points/positions.