chore(build): fix default version for robot-stack
When the git versioning system can't find a matching tag, it uses a
standard '0.0.0-dev'. When it was looking for a tag that starts with
`v*`, the default is `v0.0.0-dev`, which is then split on `v`, and so we
lost the last v in `dev`, which is invalid to python.

This situation only happens when something is building the version from
a checkout of the project without git metadata - a bare clone, a tarball
download, something like that. It will now use a correct version.