chore(build): establish an OT-3 spin of the app (#11880)
This is a new "build variant" (new terminology alert!) of the app
created when you build the app for the ot3 project. It gets a different
product name so you can install it side-by-side with the production
opentrons app, and a different app icon so you can tell it apart.

The ot3 spin, which is built when you set OPENTRONS_PROJECT=ot3 as an environment variable or make argument, uploads its builds to a different server, uses a different git tag for versioning, and looks for its updates from somewhere else. This allows us to establish parallel releases for the OT-3.

This also fixes an issue in edge wher ethe app could not run on OSX because it included an unsigned binary that was a side effect of the usb detection build.