refactor(robot-server): handle fw update errors
The firmware update process will both provide a way to statically
inspect the update state of the system with a hardware control query and
dynamically raise exceptions when the system is not updated and you try
and interact with it. While we'll protect most processes with inspection
of the update state before beginning them, we should also properly
handle the errors and present a nice json error response for them.