hack(hardware)!: hardcode left pipette presence (#12002)
After https://github.com/Opentrons/ot3-firmware/pull/544, which is
required because of hardware changes and short time to adapt, the head
is (temporarily) no longer capable of detecting the presence of
pipettes. This is a problem for the pipette QC script, which uses the
mainline pipette detection which relies on the head. Temporarily work
around this by hardcoding the python side to always return that a
pipette is present on the left mount.

This will make the script only work for a pipette on the left mount.

This may be merged into the working branch for DVT, but it MUST NOT be
merged in to edge, ever.