refactor(api): add asyncio yields to simulators
The hardware simulators all work by doing extremely minimal work in
their leaf functions. This is good and fast, except that it means that
those leaf functions often don't ever actually yield to the asyncio
loop. Tasks only yield when they hit a file io (including some asyncio
functions, like sleep(), that defer back to the loop's select call).
That means that if you simulate a long enough protocol, the task that
runs the protocol won't ever yield back to the loop, and will starve
other tasks.

To fix this, drop in a bunch of await asyncio.sleep(0) in the simulator
leaf function calls, which should defer back to the loop and fix the issue.