fix(hardware): do not track tip motor positions
For now, we should skip tracking tip motor positions so it does not accumulate with the plunger
motor positions. I made a ticket to think about how best to track this since all three motors share
the same node id and we currently store position in the `move_group_runner` by node id.