feat(hardware): support new revision values
DeviceInfo now has (or will have) revision information, but at least
some devices - DVT bootloaders - will not have, and can never have,
revision information. That means we need to tolerate messages that don't
have that data, whether they express it as 0s or as cutting off early.

Add the new data; add the handling for not having the data; and add it
within the framework of our lovely payload dataclasses, which is a
little awkward.

This needs to be merged before we start using the new data from the
firmware side, but it is safe to merge on its own because it handles
missing data.

Closes RET-1321