fix(app): show successful wifi disconnect when request pending and robot connectable
fixes the bug where the disconnect modal spins infinitely on successful disconnect. if the
disconnect request is sent when there is no wired connection, we will not receive a success response
to the request once wi-fi has disconnected. to fix, once the request is pending we check for
connectable robot health status and presume successful disconnection if the request is pending and
the robot is connectable.

closes RQA-513