fix(shared-data): add a workaround for pipette simulation
Right now, if an expected pipette name is passed in we rely on the "generation" part of the string
to determine the version. The 96 channel string did not have a "gen3" version attached and hence was
defaulting to v1.0. I'm not sure what the best path forward here is other than to better solidify
how we check that a pipette is valid, but I think we'll need to coordinate with RSS to figure it
out. For now, I am adding in this workaround so protocols can be simulated successfully.