Addition to the caution note
Adding more text to point out that user can put more than 1 liquid in a well, but if they use these methods, the best practice is to make sure the well labels reflect this.  I'm not a scientist, but it does seem like a procedure may call for additional liquids in the same well, so the note doesn't say use one liquid only. It just (I hope) conveys the idea that you ought to avoid it. @ecormany maybe give this a once over to make sure I've captured this correctly. Thanks!