fix(app): update ui when deck calibration completed (#12294)
* fix(app): update ui when deck calibration completed

Previously the UI would not update and show the new last completion date once the deck calibration
was complete. We were using `useDeckCalibrationData` to get the data and attempting to refresh data
using a polling `useCalibrationStatusQuery` call. This wasn't working as expected, so pulling the
necessary data from the result of the `useCalibrationStatusQuery` result for the deck calibrations
instead of the deck calibration data hook resolves the issue of the UI not updating like it does for
other calibrations

closes RAUT-368