docs(api): define and load liquid methods (#12251)
* docs(api): define and load liquid methods

Additions to new_labware.rst that create a new section for define_liquid and load_liquid. New section is titled "Well Labels".

Also, added an anchor tag to the "Commands" section of tutorial.rst for a cross reference link.

* Pre lunch push

Starting new drafts.

Saving in case the laptop falls into a lake or in the event of an alien invasion.

* Another set of revisions

A bigger push that includes:

- Revised section title
- Revised intro text, introduces methods in a 2nd graf after a method summary
- Removes large code sample in favor of 2 smaller snippets that focus on each method.
- Removes table that explained code. Explanations now live next to the code in respective H3 sections.
- Add a note to capture Ed's comments about volume amount validation in code (it doesn't validate).

More TK

* More revisions

This puts content in a better state. This push and draft are ready for additional review.

* One more final-ish push

Tweaks to the Note text

* Parallelism

Keeping text the same in the define an load liquid examples. One sentence ended "if you import it . . ." the other eneded "if you import your protocol ..."

the latter is better. Changing the former to match.

* Remove sentence from load_liquid section

Removing this text because the liquid object doesn't show up in the app if you don't use it. It's just a definition that sits there until used and then it will show up.

* prose pass
emphasis on beginning of protocol
capitalization and patch version in 'Opentrons App v6.3.0'
noted that all arguments are required
moved hex explanation after snippet
added exact locations where liquid info can be found in app

* code style pass
full names for first mention of methods, since they're from different classes.
ran snippets through black formatter
moved/added versionadded notices to go right after snippets

* proofread pass
🇺🇸spellings of labeling, labeled
standardize capital L for liter abbrs

* proper in-app terms

* Addition to the caution note

Adding more text to point out that user can put more than 1 liquid in a well, but if they use these methods, the best practice is to make sure the well labels reflect this.  I'm not a scientist, but it does seem like a procedure may call for additional liquids in the same well, so the note doesn't say use one liquid only. It just (I hope) conveys the idea that you ought to avoid it. @ecormany maybe give this a once over to make sure I've captured this correctly. Thanks!

* Update new_labware.rst

Adding recommended changes to last sentence from @ecormany

---------

Co-authored-by: Edward Cormany <ecormany@gmail.com>