feat(robot-server): add authentication verification
This is a proof-of-concept implementation of authentication verification. A new header is added to most endpoints to accept an `authenticationBearer` token, and a new API version (5) makes use of this token by sending it to the System Server (which must be running locally) to check whether it is an authorized token.