fix(api): allow errors in initial home z (#12287)
We saw an issue on the flex where if the estop was asserted during
startup, the axis controllers would swallow move messages instead of
sending errors - and if those move messages were homing commands with
very long timeouts, it could delay server startup indefinitely.

After https://github.com/Opentrons/ot3-firmware/pull/599 we will instead
get an error message when we send a command while estop is enabled. This
is better because we can now detect the problem, but an error occuring
at this point in initialization would prevent the robot server from
starting up, which is also bad. So let's swallow exceptions from motion
here so that the robot server doesn't stop working.

Closes RET-1342