fix(app): calibration dashboard with all tasks complete shouldnt show complete screen when visited (#12335)
The calibration dashoard has a "Calibrations Complete!" screen that should show when a user finishes
performing some number of calibrations. This screen should not show when the user visits the
calibration dashboard if all the caliibrations are already complete however. This adds a
`hasLaunchedWizard` flag that ensures that the calibrations complete screen only shows after a user
has launched one of the calibration wizards first, preventing the case were it would show upon
initial visit to the page.

closes RAUT-365