fix(app): apply LPC copy fixes and update run preview jumped step color
When a user jumps to a step from a pill in the run progress meter, it now momentarily highlights
purple. Make LPC return tip title more explicit that the same tip rack is supposed to be returned to
the slot. Fix instructions  for furthest tip/nozzle from you instead of closest.

Closes RAUT-358