refactor(hardware_control): make the gripper back paddle use the primary sensor id
All of the pipettes have their back channel (A1) as the "primary" or "S0" label. To make firmware
easier, we decided to switch the gripper ID to match the same. The calibration flow will work the
same in practice with the only difference being that the CAN message sensor ids are now flipped.