refactor: move labware hash+uri to shared-data
A lot of the places that things were importing calibration_storage were
only to access the helper functions for labware URIs and labware hashes.
Those are used in other places, and only depend on labware data, so
let's put them in the labware submodule of shared-data. This cuts
imports from ~x to ~85.