fix(api,hardware): update devices in bootloader
Currently, if we have a device in its bootloader during our network
probe, we probably don't update its firmware properly. That's for a
couple reasons:
- If it's a pipette, we don't actually know what pipette type it is,
since it won't be in attached pipettes and the bootloader can't give us
the type
- Even if it's not, we might not handle the node id properly

This commit fixes both problems.

For pipettes, we take advantage of the device subidentifier recently
introduced to the device info response payload. That subidentifier is
device-specific; for the pipettes it will be the pipette type, and for
others it isn't specified and is 0. That means that we can now find the
right pipette type for a pipette that is stuck in its bootloader. It
also means we can simplify the hardware interface by no longer requiring
the attached pipettes dict, though we only want that once we're
confident all pipettes have the new bootloader.

As a fallback for everything else, we should now explicitly handle
bootloader nodes at every stage.

Closes RQA-584