build: make sure we have annotated tags
The github actions/checkout action tries to smoothly handle tag changes
by converting annotated tags to lightweight tags when you specify an
annotated tag to checkout:
https://github.com/actions/checkout/issues/290

This breaks our version-gathering infrastructure in the case where
multiple annotated tags point to a single commit. The one that's checked
out - which is usually the most recent one, that we want the git
describe command to return - won't be returned properly because (a) it
isn't annotated and we're not passing --tags and we don't want to
because we really like having tag annotations and (b) the history is
slightly wrong since now it's a lightweight tag on the specific commit
and for that we need --contains, and we don't want to do that because we
don't want containing tags.

We can fix this by undoing the tag rewriting by forcing an update back
to the original annotated tag from the server and then checking it out.