build(app): remove unused py executables
There are a bunch of executables that get installed when we add
standalone python - pip, our scripts, other package stuff - that aren't
used at all but, on windows, have to be code signed. Since code signing
is flakey when (hypothesis) there are many things to sign, this makes
builds flakey. The hope here is that removing those extra exes means
that the signing process will become more reliable.