feat(api): expose versions of connected micros (#12470)
Implement the get_fw_version method of OT3API to return a stringified
representation of all unique firmware versions for attached
microcontrollers on the system.

This is a not-great way of exposing this data programmatically, but this
element is only used for display purposes so I think it's ok. In all
normal cases (i.e. except during update) this will only return a single
version.

Closes RCORE-731