refactor(api): ThreadManager handles async gens
Async generator functions are different from normal async functions, in the same
way that generator functions are different from normal functions. When called,
they return a special async generator object. This is a problem in
ThreadManager land, because those async generator objects need to be
interacted with via `await anext(gen)` in the same loop that created
them.

So, this commit adds a new case in thread manager for "what if you call
an async generator function in the hardware controller". In that case,
what happens is
- We return a proxy async generator function, which is defined in (and
therefore runs in) the calling loop
- We generate a future that is a coroutine running in the hardware
controller loop that captures a queue, and pulls items out of the async
generator in that loop and puts them in a queue
- Our proxy pulls stuff out of the queue and yields this

This is of course extra frustrating because all queue interaction must
take place in the same loop, which must be the loop the queue was
created in. That means that for the queue that moves generator output
from the hardware thread/loop to the calling thread/loop, it has to be
- Created in the hardware thread/loop (via run_coroutine_threadsafe and
a wrapper)
- Inserted into in the hardware thread/loop (fine, this happens in the
coroutine that drains the generator)
- Gotten from in the hardware thread/loop, so every await queue.get()
needs to happen in a run_coroutine_threadsafe()

It all works! But it's pretty annoying.