build(robot-server): add ot3-only mark handling
The robot server runs tests in a case where it's forced to be on an ot2.
This is really useful, but it leaves tests for flex-specific
functionality up a creek. With this change, if you annotate your tests
with pytests.mark.ot3_only, they just won't run in that case.