refactor(api): despecialize fw update progress
We have this getter called get_firmware_update_progress that would
return information only for mounts that were being updated, which
required some data massaging and, while, helpful for a specific use
case, (a) didn't really generalize (b) and will probably have issues
when we get the robot server handling updating firmware for things that
aren't on mounts, and in particular of pipettes that have a broken
application and jump right to the bootloader when they boot.

Making this express itself in subsystems solves that problem.