refactor(robot-server,api): fw update integration
The robot server needs to be able to perform instrument firmware
updates. The problem is that firmware updates can take a while, and the
robot server needs to have fast individual route handlers because that's
how http works - can't keep clients waiting. That means we need some
kind of logic running in the background.

The FirmwareUpdateManager provides that. It interacts with a TaskRunner
instance to run the firmware updates in the background.

The firmware updates are done with an async generator from the hardware
controller that yields out progress updates. Since that's running in one
task, and multiple other tasks (like, concurrent requests from one or
many clients) might ask for its status, we need some kind of
async-reentrant mechanism for sharing out status, which we do with a
combination of (1) an async queue for moving out of the work task
and (2) updating a "last-update" cache in sync code so that it's
async-atomic.

Co-Developed-By: Sanniti <sanniti@opentrons.com>