feat(robot-server): add flex instrument fw update
The robot server needs to be able to start, monitor, and express errors
from the update process for Flex attached instruments (like grippers and
pipettes). Let's add some new API to do that:

- POST /instruments/update {mount} starts an update, if one is
  available, and gives you an id
- GET /instruments/update/:id gives you progress information on a given
  update, if it exists

These are based on new information in the result of GET /instruments:
the current and next firmware version for the tool and whether an update
is required. Clients should not do their own version math here; updates
are held and managed by the robot server and the version fields are for
information only. Decisionmaking should be based on only the
updates-available flag.

Internally, the routes use the firmware update manager: you start a new
update process in POST /instruments/firmware and you get the status of a
running one in GET /instruments/firmware/:id. Most of this code is error
management, and we do have a couple new very specific errors, including
some new ones in the global robot_server.errors namespace that are used
all over the place and can now be centralized.

This functionality is not yet complete, in that it does not yet handle
the case where a tool is connected with a broken application firmware,
and gets stuck in its bootloader. In that case, and with the current
software stack, controlling firmware updates by mounts may not suffice,
since these aren't valid pipettes that would show up in GET
/instruments. More work needed there.

Co-Developed-By: Sanniti <sanniti@opentrons.com>