feat(api): make fw updates optional when building the hardware controller (#12413)
When the environment variable `OT3_DISABLE_FW_UPDATES` is set, when you start the OT3 repl it won't do automatic firmware updates anymore. This is helpful for specific times when people are trying to test things.

I cannot recommend enough that you do not use this capability frequently. There's a lot of behavior that depends on firmware being up to date.

Co-authored-by: Seth Foster <seth@opentrons.com>