fix(api): hardware_control: stop() in flex works
The protocol runner calls stop() after it cancels a task.
Microcontrollers may still be executing moves. If stop() then calls
update_position() before the microcontrollers are done, the request will
time out and we'll get an error.

Let's force the microcontrollers to be done by sending a stop request
and waiting for acknowledgements.

Closes RET-1353