feat(robot-server): add maintenance run resource and routes (#12529)
Adds a new resource and routes for maintenance runs that will be used by the app for instrument attach/detach/calibration flows as well as LPC flow. A maintenance run holds a protocol engine instance and is allowed to exist together with a single, idle protocol run. Only one maintenance run can exist at a time and it has lower priority than a protocol run, such that sending a play action on a protocol run will remove the existing maintenance run.

---------

Co-authored-by: tamarzanzouri <tamar.zanzouri@opentrons.com>