fix(app): lpc: fix command ordering in pick up tip
In LPC, when we picked up a tip that we'd actually use for the rest of
LPC, we'd remove the tiprack labware before moving to the trash. This is
incorrect - that labware is still physically present, since we haven't
told the user to remove it. We used to get away with it, because it used
to be that pick_up_tip ended with a z-home that got the pipette up high
enough that even if PE thought the deck was empty we wouldn't collide.

Now, though, on the Flex we removed that home because it is slow and we
don't need it anymore, and if we tell PE that the tiprack disappeared
and then tell it to move the pipette to the trash, it will just zip it
straight over to the trash... dragging along that tiprack, which is
still physically present.

To fix this, just move to the trash before we remove the labware.