fix(api): flex: do not home non-present axes
We weren't checking whether a given axis was present or not before
homing it, which is suboptimal. Now we only home the present ones.

This fixes an issue where when the HTTP API for home was called with a
mount, we'd try and home the pipette axis of the mount. This is probably
wrong, but it also should be allowed.

Did some little refactors along with this.

Closes RET-1354