fix(robot-server): cache analysis results
Recently we started getting all analyses when mounting the device page,
and this would bog down the server so much it couldn't respond to health
requests and become unhappy. While it's not clear why that happened, one
thing that's clear is that this is bad; so let's add an LRU-cache to the
analysis results gathering to try and speed up the process.

There's a couple different ways to do this and few are simple because
the primary way we get results is async, which doesn't work with the
naive lru caching in functools. What I've chosen to do here is cache the
unpickling process, which we think is the slowest part of the whole
thing.

Because this cache is keyed on the hash of the string version of the
pickled analysis, we shouldn't need explicit cache clearing. This does
mean that the process of removing the analysis data from the database
isn't cached. We may need that later.