fix(robot-server): add a cache for analyses
Getting analysis results can really bog down the robot server - in fact,
it can block the event loop such that new http requests get stalled, and
that means that apps will mark the server unhealthy.

Recently, the app started asking for analyses much more frequently, which
would cause the robot server to bog down more frequently, which would
cause the app to mark the robot unhealthy more frequently. This was an
unintended change, and reverting it fixes the robot unhealthiness; but
getting analysis results still threatens to bog down the server, so this
memcache should fix the problem on the server.

We have similar problems (and solutions) for other heavy resources like
runs and protocols; but there we do the caching with
functools.lru_cache. This is not possible for analysis results because
we get analysis results with async functions, for a very good reason; so
instead, we need a custom cache implementation.

This cache implementation uses a python dict to hold the result objects;
since there are limited places to add results, we can add things to the
cache when they're added to the database and add them to the cache when
we get things from the database. We use a deque to maintain a FIFO for
cache eviction.