refactor: refactor for testability
By splitting the implementation to add
- a public module for CompletedAnalysisStore, which has a public
interface that can be tested, with a
- public module for the memory cache, which has a public interface that
can be tested and injected into CompletedAnalysisStore

we can get back to testing only public interfaces, which is really nice.

At the same time, we can fix the race condition where get_by_protocol_id
wasn't as reentrant as it could be. Aside from that method, since this
is async code we should be access safe because the memory cache is synchronous.