fix(robot-server): memory cache analyses (#12648)
Getting analysis results can really bog down the robot server - in fact,
it can block the event loop such that new http requests get stalled, and
that means that apps will mark the server unhealthy.

Recently, the app started asking for analyses much more frequently, which
would cause the robot server to bog down more frequently, which would
cause the app to mark the robot unhealthy more frequently. This was an
unintended change, and reverting it fixes the robot unhealthiness; but
getting analysis results still threatens to bog down the server, so this
memcache should fix the problem on the server.

We have similar problems (and solutions) for other heavy resources like
runs and protocols; but there we do the caching with
functools.lru_cache. This is not possible for analysis results because
we get analysis results with async functions, for a very good reason; so
instead, we need a custom cache implementation.

This cache implementation uses a python dict to hold the result objects;
since there are limited places to add results, we can add things to the
cache when they're added to the database and add them to the cache when
we get things from the database. We use a deque to maintain a FIFO for
cache eviction.


By splitting the implementation to add
- a public module for CompletedAnalysisStore, which has a public
interface that can be tested, with a
- public module for the memory cache, which has a public interface that
can be tested and injected into CompletedAnalysisStore

we can get back to testing only public interfaces, which is really nice.

At the same time, we can fix the race condition where get_by_protocol_id
wasn't as reentrant as it could be. Aside from that method, since this
is async code we should be access safe because the memory cache is synchronous.

---------

Co-authored-by: Max Marrone <max@opentrons.com>