refactor(api,robot-server): make SubSystem exposed
We had this type called OT3SubSystem, which was a representation of the
subsystems (programmable devices, really) that could be bolted on to the
machine. It was reasonably useful but never used to its full extent.

Now, let's use it. First, we're going to want it to work on the OT-2 as
well (just for completeness) so let's remove the OT-3 part of the name
and add a way to get to it from a mount; we'll leave a temporary alias
for other code.

As an example of how we'll use it, we can make a new set of models in
the server under subsystems/.