feat(robot-server): expose instrument subsystems
Add a new element "subsystem" to the instrument responses that lets
clients link an instrument to a subsystem. The subsystem is the
representation of the programmable microcontroller on the machine, and
is the part you can update. This is an important distinction to make
because if a subsystem isn't working properly, you generally can't
identify it as an instrument; but you can always identify it as a
subsystem. The subsystem retrieved here will be useful later in some new
routes.

Note: this commit definitely won't pass tests and mypy or like work on
its own, because this is a big interconnected PR.