refactor(hardware): remove pipette lookup fallback
It's now been long enough since we got updated bootloaders that we don't
need the fallback for pipettes with bootloaders from EVT; let's remove
the fallback for looking up the pipette type in the bootloader.

Note: This commit does not stand alone because the callsites for it
aren't modified in the commit (because they're very shortly to be gone).
It's set apart so it stands out as a thing that changed.