refactor(hardware): network exposes ok status
We're going to do all this subsystem management, but a subsystem can
appear whether the device is in its application or its bootloader. We
need to expose that information somewhere, so let's start at the bottom
and use the network submodule from opentrons hardware to detect whether
the device is OK or not, and put that information in the device
information it exposes.